feat: ESPI 4.0 Schema Compliance - Phase 1: Common Domain Embeddables (#113)
* feat(phase-1): Verify and fix all common domain embeddables for ESPI 4.0 compliance

Comprehensive verification and fixes for 6 embeddable classes in domain/common:

**Critical Type Fixes (XSD Compliance)**:
- SummaryMeasurement: Changed String → UnitMultiplierKind/UnitSymbolKind enums
  - powerOfTenMultiplier: String → UnitMultiplierKind (UInt16 enum)
  - uom: String → UnitSymbolKind (UInt16 enum)
  - Added @Enumerated(EnumType.STRING) for proper JPA mapping
- ReadingInterharmonic: Changed Long → BigInteger for xs:integer compliance
  - numerator, denominator: Long → BigInteger
- DateTimeInterval: Fixed field order per XSD (duration first, then start)

**JAXB Annotations Added** (All 6 embeddables):
- @XmlAccessorType(XmlAccessType.FIELD)
- @XmlType with name, namespace, and propOrder
- @XmlElement with name and namespace on all fields
- Proper ESPI namespace: http://naesb.org/espi

**Documentation Enhancements**:
- Comprehensive Javadoc with XSD line references
- equals/hashCode via @EqualsAndHashCode (Lombok)
- Type mapping notes (BigInteger for xs:integer, enum types for UInt16)

**Database Schema Migration**:
- V1__Create_Base_Tables.sql: Changed interharmonic columns BIGINT → DECIMAL(38,0)
  - interharmonic_numerator: BIGINT → DECIMAL(38,0)
  - interharmonic_denominator: BIGINT → DECIMAL(38,0)
  - Aligns with BigInteger Java type for xs:integer XSD compliance

**Entity Updates**:
- UsageSummaryEntity.getCommodityType(): Use enum.name() instead of string
- ReadingTypeEntity: Removed temporary columnDefinition overrides

**Test Fixes**:
- LineItemRepositoryTest: Updated to use enum types, fixed DateTimeInterval order
- ReadingTypeRepositoryTest: Changed Long → BigInteger.valueOf() for interharmonic
- UsageSummaryRepositoryTest: Updated enum comparisons (UnitSymbolKind.fromValue())
- Added missing imports for UnitMultiplierKind, UnitSymbolKind, BigInteger

**Verification Results**:
- All 6 embeddables: 100% XSD compliant (espi.xsd lines 1094-1643)
- All 781 tests passing (0 failures, 0 errors)
- Type safety: 15 enum fields properly mapped
- JAXB: All embeddables ready for XML marshalling/unmarshalling

**Embeddables Verified**:
1. RationalNumber - PASS (numerator/denominator as BigInteger)
2. DateTimeInterval - PASS (duration/start with correct order)
3. SummaryMeasurement - PASS (enum types for multiplier/uom)
4. ReadingInterharmonic - PASS (BigInteger for xs:integer)
5. BillingChargeSource - PASS (JAXB annotations added)
6. LinkType - PASS (documentation enhanced)

XSD References:
- SummaryMeasurement: espi.xsd lines 1094-1129
- DateTimeInterval: espi.xsd lines 1337-1357
- RationalNumber: espi.xsd lines 1406-1418
- ReadingInterharmonic: espi.xsd lines 1419-1431
- BillingChargeSource: espi.xsd lines 1628-1643


* fix: Remove JAXB annotations from entity embeddables to resolve namespace conflicts

Removed JAXB annotations (@XmlType, @XmlAccessorType, @XmlElement) from entity
embeddable classes to fix IllegalAnnotationsException caused by duplicate XML type
names when both entity and DTO classes were loaded in the same JAXB context.

Changes:
- SummaryMeasurement.java: Removed all JAXB annotations
- BillingChargeSource.java: Removed all JAXB annotations
- RationalNumber.java: Removed all JAXB annotations
- ReadingInterharmonic.java: Removed all JAXB annotations

Architecture: Entity classes are for JPA persistence only. DTOs handle XML
marshalling with JAXB annotations. This separation prevents namespace conflicts
and follows proper DTO pattern.

DateTimeInterval and LinkType keep JAXB annotations as they have no corresponding
DTOs.

Fixes CI/CD failure in PR #113 - DataCustodianApplicationTest now passes.
All 781 openespi-common tests passing. All 3 openespi-datacustodian tests passing.

Related to #101 - Phase 1: Common Embeddables ESPI 4.0 Compliance


* refactor: Fix CustomerAgreement schema structure in customer_4.1.xsd

Changed CustomerAgreement to extend Object instead of Document, with Document
as an optional nested element. This aligns with proper schema composition pattern.

Changes:
- CustomerAgreement now extends Object (was Document)
- Added optional "document" element of type Document
- Removed XML comments for AssetContainer, OrganisationRole, WorkLocation

This change supports proper entity modeling where CustomerAgreement is a
top-level resource that composes Document information rather than inheriting
from it directly.

Related to #101 - Phase 1: Common Embeddables ESPI 4.0 Compliance


* docs: Add strict JAXB annotation guidelines to CLAUDE.md

Added critical architectural rule: JAXB annotations MUST only be applied to
DTO classes, NEVER to entity or embeddable classes.

Key points:
- Entity/Embeddable classes: JPA annotations only, NO JAXB annotations
- DTO classes: JAXB annotations only, NO JPA annotations
- If embeddable needs XML serialization: Create DTO, don't add JAXB to entity
- This rule has NO exceptions

Rationale: Prevents IllegalAnnotationsException from duplicate XML type names
when both entity and DTO classes are loaded in same JAXB context.

This documentation codifies the architectural pattern established by fixing
the JAXB namespace conflict in PR #113.

Related to #101 - Phase 1: Common Embeddables ESPI 4.0 Compliance


---------

+++ openespi-common/src/main/resources/db/migration/V1__Create_Base_Tables.sql
@@ -246,8 +246,8 @@ CREATE TABLE reading_types
     tou                     VARCHAR(50),
     uom                     VARCHAR(50),
     cpp                     VARCHAR(50),
-    interharmonic_numerator BIGINT,
-    interharmonic_denominator BIGINT,
+    interharmonic_numerator DECIMAL(38,0),
+    interharmonic_denominator DECIMAL(38,0),
     measuring_period        VARCHAR(50),
     argument_numerator      DECIMAL(38,0),
     argument_denominator    DECIMAL(38,0)
